Move tests to individual modules
