+++ src/actions/i3action.rs
@@ -41,3 +41,84 @@ impl I3ActionExt for I3Action {
         }
     }
 }
+
+#[cfg(test)]
+mod test {
+    use crate::actions::{ActionController, ActionMap, Opts};
+    use crate::test_utils::init_listener;
+    use clap::Clap;
+    use std::env;
+    use std::sync::{Arc, Mutex};
+
+    #[test]
+    /// Test the triggering of commands for the 4x2 swipe actions.
+    fn test_i3_swipe_actions() {
+        // Initialize the command line options.
+        let mut opts: Opts = Opts::parse();
+        opts.enabled_action_types = vec!["i3".to_string()];
+        opts.swipe_right_3 = vec!["i3:swipe right 3".to_string()];
+        opts.swipe_left_3 = vec!["i3:swipe left 3".to_string()];
+        opts.swipe_up_3 = vec!["i3:swipe up 3".to_string()];
+        opts.swipe_down_3 = vec!["i3:swipe down 3".to_string()];
+        opts.swipe_right_4 = vec!["i3:swipe right 4".to_string()];
+        opts.swipe_left_4 = vec!["i3:swipe left 4".to_string()];
+        opts.swipe_up_4 = vec!["i3:swipe up 4".to_string()];
+        opts.swipe_down_4 = vec!["i3:swipe down 4".to_string()];
+        opts.threshold = 5.0;
+
+        // Create the expected commands (version + 4 swipes).
+        let expected_commands = vec![
+            "swipe right 3",
+            "swipe left 3",
+            "swipe up 3",
+            "swipe down 3",
+            "swipe right 4",
+            "swipe left 4",
+            "swipe up 4",
+            "swipe down 4",
+        ];
+
+        // Create the listener and the shared storage for the commands.
+        let message_log = Arc::new(Mutex::new(vec![]));
+        init_listener(Arc::clone(&message_log));
+
+        // Trigger swipe in the 4 directions.
+        let mut action_map: ActionMap = ActionController::new(&opts);
+        action_map.populate_actions(&opts);
+        action_map.receive_end_event(&10.0, &0.0, 3);
+        action_map.receive_end_event(&-10.0, &0.0, 3);
+        action_map.receive_end_event(&0.0, &10.0, 3);
+        action_map.receive_end_event(&0.0, &-10.0, 3);
+        action_map.receive_end_event(&10.0, &0.0, 4);
+        action_map.receive_end_event(&-10.0, &0.0, 4);
+        action_map.receive_end_event(&0.0, &10.0, 4);
+        action_map.receive_end_event(&0.0, &-10.0, 4);
+
+        // Assert over the expected messages.
+        let messages = message_log.lock().unwrap();
+        assert_eq!(messages.len(), 8);
+        for (message, expected_command) in messages.iter().zip(expected_commands.iter()) {
+            assert_eq!(message, expected_command);
+        }
+    }
+
+    #[test]
+    ///Test graceful handling of unavailable i3 connection.
+    fn test_i3_not_available() {
+        // Initialize the command line options.
+        let mut opts: Opts = Opts::parse();
+        opts.enabled_action_types = vec!["i3".to_string(), "command".to_string()];
+        opts.swipe_right_3 = vec![
+            "i3:swipe right".to_string(),
+            "command:touch /tmp/swipe-right".to_string(),
+        ];
+
+        // Create the action map.
+        env::set_var("I3SOCK", "/tmp/non-existing-socket");
+        let mut action_map: ActionMap = ActionController::new(&opts);
+        action_map.populate_actions(&opts);
+
+        // Assert that only the command action is created.
+        assert_eq!(action_map.swipe_right_3.len(), 1);
+    }
+}
